Hoist PermissionSelect out of AdminPermissionsPage
The drawer Rules tab needs the same permission picker, chip dots, and
level options. Both pages now import them from `@/components`, rather
than the drawer page reaching into a sibling page for them.

Pure move: no behavior change.
